pyproject.toml
[project]
name = "cliver"
version = "0.0.4"
description = "General-purpose AI agent for your terminal — safe, controlled, and adaptable to any domain."
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "Lin Gao", email = "aoingl@gmail.com" }
]
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "click>=8.1.0",
    "rich>=13.0.0",
    "httpx>=0.24.0",
    "pydantic>=2.0.0",
    "prompt_toolkit>=3.0.0",
    "pyyaml>=6.0.0",
    "mcp>=1.2.1",
    "openai>=1.0.0",
    "langchain>=0.3.0",
    "langchain-mcp-adapters>=0.1.4",
    "langchain-openai>=0.3.18",
    "langchain-core>=0.3.63",
    "langchain-ollama>=0.3.3",
    "python-dotenv>=1.1.0",
    "jinja2>=3.1.0",
    "requests>=2.32.3",
    "json-repair>=0.52.0",
    "keyring>=25.0.0",
    "croniter>=2.0.0",
    "tomli; python_version < '3.11'"
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-cov",
    "pytest-asyncio>=1.2.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "ruff>=0.0.270",
    "black>=25.1.0",
    "pymdown-extensions>=10.16.1",
]

docs = [
    "mkdocs>=1.4.0",
    "mkdocs-material>=9.0.0",
    "mkdocs-minify-plugin>=0.2.0",
]

[project.optional-dependencies]
telegram = ["python-telegram-bot>=21.0"]
discord = ["discord.py>=2.3"]
slack = ["slack-bolt>=1.18", "slack-sdk>=3.27"]
wechat = ["aiohttp>=3.9"]
feishu = ["aiohttp>=3.9"]
api = ["aiohttp>=3.9"]
browser = ["firecrawl-py>=1.0", "playwright>=1.40"]
all = ["python-telegram-bot>=21.0", "discord.py>=2.3", "slack-bolt>=1.18", "slack-sdk>=3.27", "aiohttp>=3.9", "firecrawl-py>=1.0", "playwright>=1.40"]

[project.scripts]
cliver = "cliver.cli:cliver_main"

[project.urls]
Homepage = "https://github.com/cliver-project/CLIver"
Documentation = "https://cliver-project.github.io/CLIver/"
Repository = "https://github.com/cliver-project/CLIver"
Issues = "https://github.com/cliver-project/CLIver/issues"

[build-system]
requires = ["uv_build>=0.7,<0.8"]
build-backend = "uv_build"

[tool.uv]
package = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "B", "I"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*" = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

